Delete Makefile target if command fails
Makefile builds targets via I/O redirection (i.e. "command > file").
If the command fails, the I/O redirection will create an empty output
file with a current timestamp so it will appear to make as if it is
up-to-date when in fact it is not.  To fix this, the output file is
deleted if the command fails.